Enum rusoto_rds::CreateDBClusterError

pub enum CreateDBClusterError {
    DBClusterAlreadyExistsFault(String),
    DBClusterNotFoundFault(String),
    DBClusterParameterGroupNotFoundFault(String),
    DBClusterQuotaExceededFault(String),
    DBInstanceNotFoundFault(String),
    DBSubnetGroupDoesNotCoverEnoughAZs(String),
    DBSubnetGroupNotFoundFault(String),
    InsufficientStorageClusterCapacityFault(String),
    InvalidDBClusterStateFault(String),
    InvalidDBInstanceStateFault(String),
    InvalidDBSubnetGroupStateFault(String),
    InvalidSubnet(String),
    InvalidVPCNetworkStateFault(String),
    KMSKeyNotAccessibleFault(String),
    StorageQuotaExceededFault(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateDBCluster

Variants
DBClusterAlreadyExistsFault(String)

User already has a DB cluster with the given identifier.

DBClusterNotFoundFault(String)

DBClusterIdentifier does not refer to an existing DB cluster.

DBClusterParameterGroupNotFoundFault(String)

DBClusterParameterGroupName does not refer to an existing DB Cluster parameter group.

DBClusterQuotaExceededFault(String)

User attempted to create a new DB cluster and the user has already reached the maximum allowed DB cluster quota.

DBInstanceNotFoundFault(String)

DBInstanceIdentifier does not refer to an existing DB instance.

DBSubnetGroupDoesNotCoverEnoughAZs(String)

Subnets in the DB subnet group should cover at least two Availability Zones unless there is only one Availability Zone.

DBSubnetGroupNotFoundFault(String)

DBSubnetGroupName does not refer to an existing DB subnet group.

InsufficientStorageClusterCapacityFault(String)

There is insufficient storage available for the current action. You may be able to resolve this error by updating your subnet group to use different Availability Zones that have more storage available.

InvalidDBClusterStateFault(String)

The DB cluster is not in a valid state.

InvalidDBInstanceStateFault(String)

The specified DB instance is not in the available state.

InvalidDBSubnetGroupStateFault(String)

The DB subnet group cannot be deleted because it is in use.

InvalidSubnet(String)

The requested subnet is invalid, or multiple subnets were requested that are not all in a common VPC.

InvalidVPCNetworkStateFault(String)

DB subnet group does not cover all Availability Zones after it is created because users' change.

KMSKeyNotAccessibleFault(String)

Error accessing KMS key.

StorageQuotaExceededFault(String)

Request would result in user exceeding the allowed amount of storage available across all DB instances.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateDBClusterError

fn from_body(body: &str) -> CreateDBClusterError