Enum rusoto_rds::ModifyDBClusterError

pub enum ModifyDBClusterError {
    DBClusterAlreadyExistsFault(String),
    DBClusterNotFoundFault(String),
    DBClusterParameterGroupNotFoundFault(String),
    DBSubnetGroupNotFoundFault(String),
    InvalidDBClusterStateFault(String),
    InvalidDBInstanceStateFault(String),
    InvalidDBSecurityGroupStateFault(String),
    InvalidDBSubnetGroupStateFault(String),
    InvalidSubnet(String),
    InvalidVPCNetworkStateFault(String),
    StorageQuotaExceededFault(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by ModifyDBCluster

Variants
DBClusterAlreadyExistsFault(String)

User already has a DB cluster with the given identifier.

DBClusterNotFoundFault(String)

DBClusterIdentifier does not refer to an existing DB cluster.

DBClusterParameterGroupNotFoundFault(String)

DBClusterParameterGroupName does not refer to an existing DB Cluster parameter group.

DBSubnetGroupNotFoundFault(String)

DBSubnetGroupName does not refer to an existing DB subnet group.

InvalidDBClusterStateFault(String)

The DB cluster is not in a valid state.

InvalidDBInstanceStateFault(String)

The specified DB instance is not in the available state.

InvalidDBSecurityGroupStateFault(String)

The state of the DB security group does not allow deletion.

InvalidDBSubnetGroupStateFault(String)

The DB subnet group cannot be deleted because it is in use.

InvalidSubnet(String)

The requested subnet is invalid, or multiple subnets were requested that are not all in a common VPC.

InvalidVPCNetworkStateFault(String)

DB subnet group does not cover all Availability Zones after it is created because users' change.

StorageQuotaExceededFault(String)

Request would result in user exceeding the allowed amount of storage available across all DB instances.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ModifyDBClusterError

pub fn from_body(body: &str) -> ModifyDBClusterError