Enum rusoto_rds::ModifyDBInstanceError

pub enum ModifyDBInstanceError {
    AuthorizationNotFoundFault(String),
    CertificateNotFoundFault(String),
    DBInstanceAlreadyExistsFault(String),
    DBInstanceNotFoundFault(String),
    DBParameterGroupNotFoundFault(String),
    DBSecurityGroupNotFoundFault(String),
    DBUpgradeDependencyFailureFault(String),
    DomainNotFoundFault(String),
    InsufficientDBInstanceCapacityFault(String),
    InvalidDBInstanceStateFault(String),
    InvalidDBSecurityGroupStateFault(String),
    InvalidVPCNetworkStateFault(String),
    OptionGroupNotFoundFault(String),
    ProvisionedIopsNotAvailableInAZFault(String),
    StorageQuotaExceededFault(String),
    StorageTypeNotSupportedFault(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by ModifyDBInstance

Variants
AuthorizationNotFoundFault(String)

Specified CIDRIP or EC2 security group is not authorized for the specified DB security group.

RDS may not also be authorized via IAM to perform necessary actions on your behalf.

CertificateNotFoundFault(String)

CertificateIdentifier does not refer to an existing certificate.

DBInstanceAlreadyExistsFault(String)

User already has a DB instance with the given identifier.

DBInstanceNotFoundFault(String)

DBInstanceIdentifier does not refer to an existing DB instance.

DBParameterGroupNotFoundFault(String)

DBParameterGroupName does not refer to an existing DB parameter group.

DBSecurityGroupNotFoundFault(String)

DBSecurityGroupName does not refer to an existing DB security group.

DBUpgradeDependencyFailureFault(String)

The DB upgrade failed because a resource the DB depends on could not be modified.

DomainNotFoundFault(String)

Domain does not refer to an existing Active Directory Domain.

InsufficientDBInstanceCapacityFault(String)

Specified DB instance class is not available in the specified Availability Zone.

InvalidDBInstanceStateFault(String)

The specified DB instance is not in the available state.

InvalidDBSecurityGroupStateFault(String)

The state of the DB security group does not allow deletion.

InvalidVPCNetworkStateFault(String)

DB subnet group does not cover all Availability Zones after it is created because users' change.

OptionGroupNotFoundFault(String)

The specified option group could not be found.

ProvisionedIopsNotAvailableInAZFault(String)

Provisioned IOPS not available in the specified Availability Zone.

StorageQuotaExceededFault(String)

Request would result in user exceeding the allowed amount of storage available across all DB instances.

StorageTypeNotSupportedFault(String)

StorageType specified cannot be associated with the DB Instance.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ModifyDBInstanceError

pub fn from_body(body: &str) -> ModifyDBInstanceError