Struct rusoto_rds::CreateDBClusterMessage

pub struct CreateDBClusterMessage {
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub db_cluster_identifier: String,
    pub db_cluster_parameter_group_name: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub database_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub kms_key_id: Option<String>,
    pub master_user_password: Option<String>,
    pub master_username: Option<String>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub pre_signed_url: Option<String>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_source_identifier: Option<String>,
    pub storage_encrypted: Option<bool>,
    pub tags: Option<Vec<Tag>>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
availability_zones: Option<Vec<String>>

A list of EC2 Availability Zones that instances in the DB cluster can be created in. For information on regions and Availability Zones, see Regions and Availability Zones.

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

character_set_name: Option<String>

A value that indicates that the DB cluster should be associated with the specified CharacterSet.

db_cluster_identifier: String

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

db_cluster_parameter_group_name: Option<String>

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, default.aurora5.6 will be used.

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB cluster.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

database_name: Option<String>

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon RDS will not create a database in the DB cluster you are creating.

enable_iam_database_authentication: Option<bool>

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

engine: String

The name of the database engine to be used for this DB cluster.

Valid Values: aurora

engine_version: Option<String>

The version number of the database engine to use.

Aurora

Example: 5.6.10a

kms_key_id: Option<String>

The KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

If you create a Read Replica of an encrypted DB cluster in another region, you must set KmsKeyId to a KMS key ID that is valid in the destination region. This key is used to encrypt the Read Replica in that region.

master_user_password: Option<String>

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

master_username: Option<String>

The name of the master user for the DB cluster.

Constraints:

Must be 1 to 16 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

option_group_name: Option<String>

A value that indicates that the DB cluster should be associated with the specified option group.

Permanent options cannot be removed from an option group. The option group cannot be removed from a DB cluster once it is associated with a DB cluster.

port: Option<i64>

The port number on which the instances in the DB cluster accept connections.

Default: 3306

pre_signed_url: Option<String>

A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source region where the DB cluster will be replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster.

The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source region that contains the encrypted DB cluster to be copied.

The pre-signed URL request must contain the following parameter values:

KmsKeyId - The KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination region. This should refer to the same KMS key for both the CreateDBCluster action that is called in the destination region, and the action contained in the pre-signed URL.

DestinationRegion - The name of the region that Aurora Read Replica will be created in.

ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source region. For example, if you are copying an encrypted DB cluster from the us-west-2 region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (AWS Signature Version 4) and Signature Version 4 Signing Process.

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

Default: A 30-minute window selected at random from an 8-hour block of time per region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Times should be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

replication_source_identifier: Option<String>

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

storage_encrypted: Option<bool>

Specifies whether the DB cluster is encrypted.

tags: Option<Vec<Tag>> vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB cluster.