Struct rusoto_rds::CreateDBInstanceMessage

pub struct CreateDBInstanceMessage {
    pub allocated_storage: Option<i64>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_cluster_identifier: Option<String>,
    pub db_instance_class: String,
    pub db_instance_identifier: String,
    pub db_name: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub db_subnet_group_name: Option<String>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub license_model: Option<String>,
    pub master_user_password: Option<String>,
    pub master_username: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub promotion_tier: Option<i64>,
    pub publicly_accessible: Option<bool>,
    pub storage_encrypted: Option<bool>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub timezone: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
allocated_storage: Option<i64>

The amount of storage (in gigabytes) to be initially allocated for the database instance.

Type: Integer

Amazon Aurora

Not applicable. Aurora cluster volumes automatically grow as the amount of data in your database increases, though you are only charged for the space that you use in an Aurora cluster volume.

MySQL

Constraints: Must be an integer from 5 to 6144.

MariaDB

Constraints: Must be an integer from 5 to 6144.

PostgreSQL

Constraints: Must be an integer from 5 to 6144.

Oracle

Constraints: Must be an integer from 10 to 6144.

SQL Server

Constraints: Must be an integer from 200 to 4096 (Standard Edition and Enterprise Edition) or from 20 to 4096 (Express Edition and Web Edition)

auto_minor_version_upgrade: Option<bool>

Indicates that minor engine upgrades will be applied automatically to the DB instance during the maintenance window.

Default: true

availability_zone: Option<String>

The EC2 Availability Zone that the database instance will be created in. For information on regions and Availability Zones, see Regions and Availability Zones.

Default: A random, system-chosen Availability Zone in the endpoint's region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter cannot be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same region as the current endpoint.

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

character_set_name: Option<String>

For supported engines, indicates that the DB instance should be associated with the specified CharacterSet.

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the DB instance to snapshots of the DB instance; otherwise false. The default is false.

db_cluster_identifier: Option<String>

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

db_instance_class: String

The compute and memory capacity of the DB instance. Note that not all instance classes are available in all regions for all DB engines.

Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.xlarge |db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large

db_instance_identifier: String

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens (1 to 15 for SQL Server).

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use.

Type: String

MySQL

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 alphanumeric characters

Cannot be a word reserved by the specified database engine

MariaDB

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 alphanumeric characters

Cannot be a word reserved by the specified database engine

PostgreSQL

The name of the database to create when the DB instance is created. If this parameter is not specified, the default "postgres" database is created in the DB instance.

Constraints:

Must contain 1 to 63 alphanumeric characters

Must begin with a letter or an underscore. Subsequent characters can be letters, underscores, or digits (0-9).

Cannot be a word reserved by the specified database engine

Oracle

The Oracle System ID (SID) of the created DB instance. If you specify null, the default value ORCL is used. You can't specify the string NULL, or any other reserved word, for DBName.

Default: ORCL

Constraints:

Cannot be longer than 8 characters

SQL Server

Not applicable. Must be null.

Amazon Aurora

The name of the database to create when the primary instance of the DB cluster is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 alphanumeric characters

Cannot be a word reserved by the specified database engine

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine will be used.

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

db_security_groups: Option<Vec<String>>

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

domain: Option<String>

Specify the Active Directory Domain to create the instance in.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_iam_database_authentication: Option<bool>

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false.

You can enable IAM database authentication for the following database engines:

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Default: false

engine: String

The name of the database engine to be used for this instance.

Not every database engine is available for every AWS region.

Valid Values:

aurora

mariadb

mysql

oracle-ee

oracle-se2

oracle-se1

oracle-se

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

engine_version: Option<String>

The version number of the database engine to use.

The following are the database engines and major and minor versions that are available with Amazon RDS. Not every database engine is available for every AWS region.

Amazon Aurora

Version 5.6 (available in these AWS regions: ap-northeast-1, ap-northeast-2, ap-south-1, ap-southeast-2, eu-west-1, us-east-1, us-east-2, us-west-2): 5.6.10a

MariaDB

10.1.19 (supported in all AWS regions)

10.1.14 (supported in all regions except us-east-2)

10.0.28 (supported in all AWS regions)

10.0.24 (supported in all AWS regions)

10.0.17 (supported in all regions except us-east-2, ca-central-1, eu-west-2)

Microsoft SQL Server 2016

13.00.4422.0.v1 (supported for all editions, and all AWS regions)

13.00.2164.0.v1 (supported for all editions, and all AWS regions)

Microsoft SQL Server 2014

12.00.5546.0.v1 (supported for all editions, and all AWS regions)

12.00.5000.0.v1 (supported for all editions, and all AWS regions)

12.00.4422.0.v1 (supported for all editions except Enterprise Edition, and all AWS regions except ca-central-1 and eu-west-2)

Microsoft SQL Server 2012

11.00.6594.0.v1 (supported for all editions, and all AWS regions)

11.00.6020.0.v1 (supported for all editions, and all AWS regions)

11.00.5058.0.v1 (supported for all editions, and all AWS regions except us-east-2, ca-central-1, and eu-west-2)

11.00.2100.60.v1 (supported for all editions, and all AWS regions except us-east-2, ca-central-1, and eu-west-2)

Microsoft SQL Server 2008 R2

10.50.6529.0.v1 (supported for all editions, and all AWS regions except us-east-2, ca-central-1, and eu-west-2)

10.50.6000.34.v1 (supported for all editions, and all AWS regions except us-east-2, ca-central-1, and eu-west-2)

10.50.2789.0.v1 (supported for all editions, and all AWS regions except us-east-2, ca-central-1, and eu-west-2)

MySQL

5.7.17 (supported in all AWS regions)

5.7.16 (supported in all AWS regions)

5.7.11 (supported in all AWS regions)

5.7.10 (supported in all regions except us-east-2, ca-central-1, eu-west-2)

5.6.35 (supported in all AWS regions)

5.6.34 (supported in all AWS regions)

5.6.29 (supported in all AWS regions)

5.6.27 (supported in all regions except us-east-2, ca-central-1, eu-west-2)

5.6.23 (supported in all regions except us-east-2, ap-south-1, ca-central-1, eu-west-2)

5.6.22 (supported in all regions except us-east-2, ap-south-1, ap-northeast-2, ca-central-1, eu-west-2)

5.6.21b (supported in all regions except us-east-2, ap-south-1, ap-northeast-2, ca-central-1, eu-west-2)

5.6.21 (supported in all regions except us-east-2, ap-south-1, ap-northeast-2, ca-central-1, eu-west-2)

5.6.19b (supported in all regions except us-east-2, ap-south-1, ap-northeast-2, ca-central-1, eu-west-2)

5.6.19a (supported in all regions except us-east-2, ap-south-1, ap-northeast-2, ca-central-1, eu-west-2)

5.5.54 (supported in all AWS regions)

5.5.53 (supported in all AWS regions)

5.5.46 (supported in all AWS regions)

Oracle 12c

12.1.0.2.v8 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v7 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v6 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v5 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v4 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v3 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v2 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v1 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

Oracle 11g

11.2.0.4.v12 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v11 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v10 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v9 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v8 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v7 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v6 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v5 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v4 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v3 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v1 (supported for EE, SE1, and SE, in all AWS regions)

PostgreSQL

Version 9.6.x: 9.6.1 | 9.6.2

Version 9.5.x: 9.5.6 | 9.5.4 | 9.5.2

Version 9.4.x: 9.4.11 | 9.4.9 | 9.4.7

Version 9.3.x: 9.3.16 | 9.3.14 | 9.3.12

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be a multiple between 3 and 10 of the storage amount for the DB instance. Must also be an integer multiple of 1000. For example, if the size of your DB instance is 500 GB, then your Iops value can be 2000, 3000, 4000, or 5000.

kms_key_id: Option<String>

The KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same AWS account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

license_model: Option<String>

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

master_user_password: Option<String>

The password for the master database user. Can be any printable ASCII character except "/", """, or "@".

Amazon Aurora

Not applicable. You specify the password for the master database user when you create your DB cluster.

MariaDB

Constraints: Must contain from 8 to 41 characters.

Microsoft SQL Server

Constraints: Must contain from 8 to 128 characters.

MySQL

Constraints: Must contain from 8 to 41 characters.

Oracle

Constraints: Must contain from 8 to 30 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

master_username: Option<String>

The name for the master database user.

Amazon Aurora

Not applicable. You specify the name for the master database user when you create your DB cluster.

MariaDB

Constraints:

Must be 1 to 16 alphanumeric characters.

Cannot be a reserved word for the chosen database engine.

Microsoft SQL Server

Constraints:

Must be 1 to 128 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

MySQL

Constraints:

Must be 1 to 16 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

Oracle

Constraints:

Must be 1 to 30 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

PostgreSQL

Constraints:

Must be 1 to 63 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to Setting Up and Enabling Enhanced Monitoring.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

option_group_name: Option<String>

Indicates that the DB instance should be associated with the specified option group.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

port: Option<i64>

The port number on which the database accepts connections.

MySQL

Default: 3306

Valid Values: 1150-65535

Type: Integer

MariaDB

Default: 3306

Valid Values: 1150-65535

Type: Integer

PostgreSQL

Default: 5432

Valid Values: 1150-65535

Type: Integer

Oracle

Default: 1521

Valid Values: 1150-65535

SQL Server

Default: 1433

Valid Values: 1150-65535 except for 1434, 3389, 47001, 49152, and 49152 through 49156.

Amazon Aurora

Default: 3306

Valid Values: 1150-65535

Type: Integer

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. For more information, see DB Instance Backups.

Default: A 30-minute window selected at random from an 8-hour block of time per region. To see the time blocks available, see Adjusting the Preferred DB Instance Maintenance Window.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Times should be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see DB Instance Maintenance.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

promotion_tier: Option<i64>

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster.

Default: 1

Valid Values: 0 - 15

publicly_accessible: Option<bool>

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC: true

VPC: false

If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

storage_encrypted: Option<bool>

Specifies whether the DB instance is encrypted.

Default: false

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

tags: Option<Vec<Tag>> tde_credential_arn: Option<String>

The ARN from the Key Store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the Key Store in order to access the device.

timezone: Option<String>

The time zone of the DB instance. The time zone parameter is currently supported only by Microsoft SQL Server.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.