Struct rusoto_rds::CreateDBInstanceReadReplicaMessage

pub struct CreateDBInstanceReadReplicaMessage {
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_instance_class: Option<String>,
    pub db_instance_identifier: String,
    pub db_subnet_group_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub pre_signed_url: Option<String>,
    pub publicly_accessible: Option<bool>,
    pub source_db_instance_identifier: String,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
}

Fields
auto_minor_version_upgrade: Option<bool>

Indicates that minor engine upgrades will be applied automatically to the Read Replica during the maintenance window.

Default: Inherits from the source DB instance

availability_zone: Option<String>

The Amazon EC2 Availability Zone that the Read Replica will be created in.

Default: A random, system-chosen Availability Zone in the endpoint's region.

Example: us-east-1d

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the Read Replica to snapshots of the Read Replica; otherwise false. The default is false.

db_instance_class: Option<String>

The compute and memory capacity of the Read Replica. Note that not all instance classes are available in all regions for all DB engines.

Valid Values: db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.xlarge |db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large

Default: Inherits from the source DB instance.

db_instance_identifier: String

The DB instance identifier of the Read Replica. This identifier is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

db_subnet_group_name: Option<String>

Specifies a DB subnet group for the DB instance. The new DB instance will be created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance is not created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another region.

The specified DB subnet group must be in the same region in which the operation is running.

All Read Replicas in one region that are created from the same source DB instance must either:>

Specify DB subnet groups from the same VPC. All these Read Replicas will be created in the same VPC.

Not specify a DB subnet group. All these Read Replicas will be created outside of any VPC.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

enable_iam_database_authentication: Option<bool>

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false.

You can enable IAM database authentication for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher.

Default: false

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

kms_key_id: Option<String>

The AWS KMS key ID for an encrypted Read Replica. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

If you create an unencrypted Read Replica and specify a value for the KmsKeyId parameter, Amazon RDS encrypts the target Read Replica using the specified KMS encryption key.

If you create an encrypted Read Replica from your AWS account, you can specify a value for KmsKeyId to encrypt the Read Replica with a new KMS encryption key. If you don't specify a value for KmsKeyId, then the Read Replica is encrypted with the same KMS key as the source DB instance.

If you create an encrypted Read Replica in a different AWS region, then you must specify a KMS key for the destination AWS region. KMS encryption keys are specific to the region that they are created in, and you cannot use encryption keys from one region in another region.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the Read Replica. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

option_group_name: Option<String>

The option group the DB instance will be associated with. If omitted, the default option group for the engine specified will be used.

port: Option<i64>

The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

pre_signed_url: Option<String>

The URL that contains a Signature Version 4 signed request for the CreateDBInstanceReadReplica API action in the AWS region that contains the source DB instance. The PreSignedUrl parameter must be used when encrypting a Read Replica from another AWS region.

The presigned URL must be a valid request for the CreateDBInstanceReadReplica API action that can be executed in the source region that contains the encrypted DB instance. The presigned URL request must contain the following parameter values:

DestinationRegion - The AWS Region that the Read Replica is created in. This region is the same one where the CreateDBInstanceReadReplica action is called that contains this presigned URL.

For example, if you create an encrypted Read Replica in the us-east-1 region, and the source DB instance is in the west-2 region, then you call the CreateDBInstanceReadReplica action in the us-east-1 region and provide a presigned URL that contains a call to the CreateDBInstanceReadReplica action in the us-west-2 region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 region.

KmsKeyId - The KMS key identifier for the key to use to encrypt the Read Replica in the destination region. This is the same identifier for both the CreateDBInstanceReadReplica action that is called in the destination region, and the action contained in the presigned URL.

SourceDBInstanceIdentifier - The DB instance identifier for the encrypted Read Replica to be created. This identifier must be in the Amazon Resource Name (ARN) format for the source region. For example, if you create an encrypted Read Replica from a DB instance in the us-west-2 region, then your SourceDBInstanceIdentifier would look like this example: arn:aws:rds:us-west-2:123456789012:instance:mysql-instance1-instance-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (AWS Signature Version 4) and Signature Version 4 Signing Process.

publicly_accessible: Option<bool>

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC:true

VPC:false

If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

source_db_instance_identifier: String

The identifier of the DB instance that will act as the source for the Read Replica. Each DB instance can have up to five Read Replicas.

Constraints:

Must be the identifier of an existing MySQL, MariaDB, or PostgreSQL DB instance.

Can specify a DB instance that is a MySQL Read Replica only if the source is running MySQL 5.6.

Can specify a DB instance that is a PostgreSQL DB instance only if the source is running PostgreSQL 9.3.5 or later.

The specified DB instance must have automatic backups enabled, its backup retention period must be greater than 0.

If the source DB instance is in the same region as the Read Replica, specify a valid DB instance identifier.

If the source DB instance is in a different region than the Read Replica, specify a valid DB instance ARN. For more information, go to Constructing a Amazon RDS Amazon Resource Name (ARN).

storage_type: Option<String>

Specifies the storage type to be associated with the Read Replica.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

tags: Option<Vec<Tag>>