Struct rusoto_rds::DBCluster

pub struct DBCluster {
    pub allocated_storage: Option<i64>,
    pub associated_roles: Option<Vec<DBClusterRole>>,
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub clone_group_id: Option<String>,
    pub cluster_create_time: Option<String>,
    pub db_cluster_arn: Option<String>,
    pub db_cluster_identifier: Option<String>,
    pub db_cluster_members: Option<Vec<DBClusterMember>>,
    pub db_cluster_option_group_memberships: Option<Vec<DBClusterOptionGroupStatus>>,
    pub db_cluster_parameter_group: Option<String>,
    pub db_subnet_group: Option<String>,
    pub database_name: Option<String>,
    pub db_cluster_resource_id: Option<String>,
    pub earliest_restorable_time: Option<String>,
    pub endpoint: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub hosted_zone_id: Option<String>,
    pub iam_database_authentication_enabled: Option<bool>,
    pub kms_key_id: Option<String>,
    pub latest_restorable_time: Option<String>,
    pub master_username: Option<String>,
    pub multi_az: Option<bool>,
    pub percent_progress: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub read_replica_identifiers: Option<Vec<String>>,
    pub reader_endpoint: Option<String>,
    pub replication_source_identifier: Option<String>,
    pub status: Option<String>,
    pub storage_encrypted: Option<bool>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
}

Contains the result of a successful invocation of the following actions:

CreateDBCluster

DeleteDBCluster

FailoverDBCluster

ModifyDBCluster

RestoreDBClusterFromSnapshot

RestoreDBClusterToPointInTime

This data type is used as a response element in the DescribeDBClusters action.

Fields
allocated_storage: Option<i64>

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gigabytes (GB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size is not fixed, but instead automatically adjusts as needed.

associated_roles: Option<Vec<DBClusterRole>>

Provides a list of the AWS Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other AWS services on your behalf.

availability_zones: Option<Vec<String>>

Provides the list of EC2 Availability Zones that instances in the DB cluster can be created in.

backup_retention_period: Option<i64>

Specifies the number of days for which automatic DB snapshots are retained.

character_set_name: Option<String>

If present, specifies the name of the character set that this cluster is associated with.

clone_group_id: Option<String>

Identifies the clone group to which the DB cluster is associated.

cluster_create_time: Option<String>

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

db_cluster_arn: Option<String>

The Amazon Resource Name (ARN) for the DB cluster.

db_cluster_identifier: Option<String>

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

db_cluster_members: Option<Vec<DBClusterMember>>

Provides the list of instances that make up the DB cluster.

db_cluster_option_group_memberships: Option<Vec<DBClusterOptionGroupStatus>>

Provides the list of option group memberships for this DB cluster.

db_cluster_parameter_group: Option<String>

Specifies the name of the DB cluster parameter group for the DB cluster.

db_subnet_group: Option<String>

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

database_name: Option<String>

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

db_cluster_resource_id: Option<String>

The region-unique, immutable identifier for the DB cluster. This identifier is found in AWS CloudTrail log entries whenever the KMS key for the DB cluster is accessed.

earliest_restorable_time: Option<String>

Specifies the earliest time to which a database can be restored with point-in-time restore.

endpoint: Option<String>

Specifies the connection endpoint for the primary instance of the DB cluster.

engine: Option<String>

Provides the name of the database engine to be used for this DB cluster.

engine_version: Option<String>

Indicates the database engine version.

hosted_zone_id: Option<String>

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

iam_database_authentication_enabled: Option<bool>

True if mapping of AWS Identity and Access Management (IAM) accounts to database accounts is enabled; otherwise false.

kms_key_id: Option<String>

If StorageEncrypted is true, the KMS key identifier for the encrypted DB cluster.

latest_restorable_time: Option<String>

Specifies the latest time to which a database can be restored with point-in-time restore.

master_username: Option<String>

Contains the master username for the DB cluster.

multi_az: Option<bool>

Specifies whether the DB cluster has instances in multiple Availability Zones.

percent_progress: Option<String>

Specifies the progress of the operation as a percentage.

port: Option<i64>

Specifies the port that the database engine is listening on.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

read_replica_identifiers: Option<Vec<String>>

Contains one or more identifiers of the Read Replicas associated with this DB cluster.

reader_endpoint: Option<String>

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection will be dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

replication_source_identifier: Option<String>

Contains the identifier of the source DB cluster if this DB cluster is a Read Replica.

status: Option<String>

Specifies the current state of this DB cluster.

storage_encrypted: Option<bool>

Specifies whether the DB cluster is encrypted.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of VPC security groups that the DB cluster belongs to.