Struct rusoto_rds::DBInstance

pub struct DBInstance {
    pub allocated_storage: Option<i64>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub backup_retention_period: Option<i64>,
    pub ca_certificate_identifier: Option<String>,
    pub character_set_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_cluster_identifier: Option<String>,
    pub db_instance_arn: Option<String>,
    pub db_instance_class: Option<String>,
    pub db_instance_identifier: Option<String>,
    pub db_instance_status: Option<String>,
    pub db_name: Option<String>,
    pub db_parameter_groups: Option<Vec<DBParameterGroupStatus>>,
    pub db_security_groups: Option<Vec<DBSecurityGroupMembership>>,
    pub db_subnet_group: Option<DBSubnetGroup>,
    pub db_instance_port: Option<i64>,
    pub dbi_resource_id: Option<String>,
    pub domain_memberships: Option<Vec<DomainMembership>>,
    pub endpoint: Option<Endpoint>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub enhanced_monitoring_resource_arn: Option<String>,
    pub iam_database_authentication_enabled: Option<bool>,
    pub instance_create_time: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub latest_restorable_time: Option<String>,
    pub license_model: Option<String>,
    pub master_username: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_memberships: Option<Vec<OptionGroupMembership>>,
    pub pending_modified_values: Option<PendingModifiedValues>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub promotion_tier: Option<i64>,
    pub publicly_accessible: Option<bool>,
    pub read_replica_db_cluster_identifiers: Option<Vec<String>>,
    pub read_replica_db_instance_identifiers: Option<Vec<String>>,
    pub read_replica_source_db_instance_identifier: Option<String>,
    pub secondary_availability_zone: Option<String>,
    pub status_infos: Option<Vec<DBInstanceStatusInfo>>,
    pub storage_encrypted: Option<bool>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub timezone: Option<String>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
}

Contains the result of a successful invocation of the following actions:

CreateDBInstance

DeleteDBInstance

ModifyDBInstance

StopDBInstance

StartDBInstance

This data type is used as a response element in the DescribeDBInstances action.

Fields
allocated_storage: Option<i64>

Specifies the allocated storage size specified in gigabytes.

auto_minor_version_upgrade: Option<bool>

Indicates that minor version patches are applied automatically.

availability_zone: Option<String>

Specifies the name of the Availability Zone the DB instance is located in.

backup_retention_period: Option<i64>

Specifies the number of days for which automatic DB snapshots are retained.

ca_certificate_identifier: Option<String>

The identifier of the CA certificate for this DB instance.

character_set_name: Option<String>

If present, specifies the name of the character set that this instance is associated with.

copy_tags_to_snapshot: Option<bool>

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

db_cluster_identifier: Option<String>

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

db_instance_arn: Option<String>

The Amazon Resource Name (ARN) for the DB instance.

db_instance_class: Option<String>

Contains the name of the compute and memory capacity class of the DB instance.

db_instance_identifier: Option<String>

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

db_instance_status: Option<String>

Specifies the current state of this database.

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use. For example, this value returns MySQL, MariaDB, or PostgreSQL information when returning values from CreateDBInstanceReadReplica since Read Replicas are only supported for these engines.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

db_parameter_groups: Option<Vec<DBParameterGroupStatus>>

Provides the list of DB parameter groups applied to this DB instance.

db_security_groups: Option<Vec<DBSecurityGroupMembership>>

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

db_subnet_group: Option<DBSubnetGroup>

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

db_instance_port: Option<i64>

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

dbi_resource_id: Option<String>

The region-unique, immutable identifier for the DB instance. This identifier is found in AWS CloudTrail log entries whenever the KMS key for the DB instance is accessed.

domain_memberships: Option<Vec<DomainMembership>>

The Active Directory Domain membership records associated with the DB instance.

endpoint: Option<Endpoint>

Specifies the connection endpoint.

engine: Option<String>

Provides the name of the database engine to be used for this DB instance.

engine_version: Option<String>

Indicates the database engine version.

enhanced_monitoring_resource_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

iam_database_authentication_enabled: Option<bool>

True if mapping of AWS Identity and Access Management (IAM) accounts to database accounts is enabled; otherwise false.

IAM database authentication can be enabled for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher. To enable IAM database authentication for Aurora, see DBCluster Type.

instance_create_time: Option<String>

Provides the date and time the DB instance was created.

iops: Option<i64>

Specifies the Provisioned IOPS (I/O operations per second) value.

kms_key_id: Option<String>

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

latest_restorable_time: Option<String>

Specifies the latest time to which a database can be restored with point-in-time restore.

license_model: Option<String>

License model information for this DB instance.

master_username: Option<String>

Contains the master username for the DB instance.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send Enhanced Monitoring metrics to CloudWatch Logs.

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment.

option_group_memberships: Option<Vec<OptionGroupMembership>>

Provides the list of option group memberships for this DB instance.

pending_modified_values: Option<PendingModifiedValues>

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

promotion_tier: Option<i64>

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster.

publicly_accessible: Option<bool>

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC:true

VPC:false

If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

read_replica_db_cluster_identifiers: Option<Vec<String>>

Contains one or more identifiers of Aurora DB clusters that are Read Replicas of this DB instance.

read_replica_db_instance_identifiers: Option<Vec<String>>

Contains one or more identifiers of the Read Replicas associated with this DB instance.

read_replica_source_db_instance_identifier: Option<String>

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

secondary_availability_zone: Option<String>

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

status_infos: Option<Vec<DBInstanceStatusInfo>>

The status of a Read Replica. If the instance is not a Read Replica, this will be blank.

storage_encrypted: Option<bool>

Specifies whether the DB instance is encrypted.

storage_type: Option<String>

Specifies the storage type associated with DB instance.

tde_credential_arn: Option<String>

The ARN from the key store with which the instance is associated for TDE encryption.

timezone: Option<String>

The time zone of the DB instance. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of VPC security group elements that the DB instance belongs to.