Struct rusoto_rds::DBSnapshot

pub struct DBSnapshot {
    pub allocated_storage: Option<i64>,
    pub availability_zone: Option<String>,
    pub db_instance_identifier: Option<String>,
    pub db_snapshot_arn: Option<String>,
    pub db_snapshot_identifier: Option<String>,
    pub encrypted: Option<bool>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub iam_database_authentication_enabled: Option<bool>,
    pub instance_create_time: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub license_model: Option<String>,
    pub master_username: Option<String>,
    pub option_group_name: Option<String>,
    pub percent_progress: Option<i64>,
    pub port: Option<i64>,
    pub snapshot_create_time: Option<String>,
    pub snapshot_type: Option<String>,
    pub source_db_snapshot_identifier: Option<String>,
    pub source_region: Option<String>,
    pub status: Option<String>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub timezone: Option<String>,
    pub vpc_id: Option<String>,
}

Contains the result of a successful invocation of the following actions:

CreateDBSnapshot

DeleteDBSnapshot

This data type is used as a response element in the DescribeDBSnapshots action.

Fields
allocated_storage: Option<i64>

Specifies the allocated storage size in gigabytes (GB).

availability_zone: Option<String>

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

db_instance_identifier: Option<String>

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

db_snapshot_arn: Option<String>

The Amazon Resource Name (ARN) for the DB snapshot.

db_snapshot_identifier: Option<String>

Specifies the identifier for the DB snapshot.

encrypted: Option<bool>

Specifies whether the DB snapshot is encrypted.

engine: Option<String>

Specifies the name of the database engine.

engine_version: Option<String>

Specifies the version of the database engine.

iam_database_authentication_enabled: Option<bool>

True if mapping of AWS Identity and Access Management (IAM) accounts to database accounts is enabled; otherwise false.

instance_create_time: Option<String>

Specifies the time when the snapshot was taken, in Universal Coordinated Time (UTC).

iops: Option<i64>

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

kms_key_id: Option<String>

If Encrypted is true, the KMS key identifier for the encrypted DB snapshot.

license_model: Option<String>

License model information for the restored DB instance.

master_username: Option<String>

Provides the master username for the DB snapshot.

option_group_name: Option<String>

Provides the option group name for the DB snapshot.

percent_progress: Option<i64>

The percentage of the estimated data that has been transferred.

port: Option<i64>

Specifies the port that the database engine was listening on at the time of the snapshot.

snapshot_create_time: Option<String>

Provides the time when the snapshot was taken, in Universal Coordinated Time (UTC).

snapshot_type: Option<String>

Provides the type of the DB snapshot.

source_db_snapshot_identifier: Option<String>

The DB snapshot Arn that the DB snapshot was copied from. It only has value in case of cross customer or cross region copy.

source_region: Option<String>

The region that the DB snapshot was created in or copied from.

status: Option<String>

Specifies the status of this DB snapshot.

storage_type: Option<String>

Specifies the storage type associated with DB snapshot.

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

timezone: Option<String>

The time zone of the DB snapshot. In most cases, the Timezone element is empty. Timezone content appears only for snapshots taken from Microsoft SQL Server DB instances that were created with a time zone specified.

vpc_id: Option<String>

Provides the VPC ID associated with the DB snapshot.