Struct rusoto_rds::OptionGroupOption

pub struct OptionGroupOption {
    pub default_port: Option<i64>,
    pub description: Option<String>,
    pub engine_name: Option<String>,
    pub major_engine_version: Option<String>,
    pub minimum_required_minor_engine_version: Option<String>,
    pub name: Option<String>,
    pub option_group_option_settings: Option<Vec<OptionGroupOptionSetting>>,
    pub option_group_option_versions: Option<Vec<OptionVersion>>,
    pub options_conflicts_with: Option<Vec<String>>,
    pub options_depended_on: Option<Vec<String>>,
    pub permanent: Option<bool>,
    pub persistent: Option<bool>,
    pub port_required: Option<bool>,
}

Available option.

Fields
default_port: Option<i64>

If the option requires a port, specifies the default port for the option.

description: Option<String>

The description of the option.

engine_name: Option<String>

The name of the engine that this option can be applied to.

major_engine_version: Option<String>

Indicates the major engine version that the option is available for.

minimum_required_minor_engine_version: Option<String>

The minimum required engine version for the option to be applied.

name: Option<String>

The name of the option.

option_group_option_settings: Option<Vec<OptionGroupOptionSetting>>

The option settings that are available (and the default value) for each option in an option group.

option_group_option_versions: Option<Vec<OptionVersion>>

The versions that are available for the option.

options_conflicts_with: Option<Vec<String>>

The options that conflict with this option.

options_depended_on: Option<Vec<String>>

The options that are prerequisites for this option.

permanent: Option<bool>

Permanent options can never be removed from an option group. An option group containing a permanent option can't be removed from a DB instance.

persistent: Option<bool>

Persistent options can't be removed from an option group while DB instances are associated with the option group. If you disassociate all DB instances from the option group, your can remove the persistent option from the option group.

port_required: Option<bool>

Specifies whether the option requires a port.