Struct rusoto_rds::OrderableDBInstanceOption

pub struct OrderableDBInstanceOption {
    pub availability_zones: Option<Vec<AvailabilityZone>>,
    pub db_instance_class: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub license_model: Option<String>,
    pub max_iops_per_db_instance: Option<i64>,
    pub max_iops_per_gib: Option<f64>,
    pub max_storage_size: Option<i64>,
    pub min_iops_per_db_instance: Option<i64>,
    pub min_iops_per_gib: Option<f64>,
    pub min_storage_size: Option<i64>,
    pub multi_az_capable: Option<bool>,
    pub read_replica_capable: Option<bool>,
    pub storage_type: Option<String>,
    pub supports_enhanced_monitoring: Option<bool>,
    pub supports_iam_database_authentication: Option<bool>,
    pub supports_iops: Option<bool>,
    pub supports_performance_insights: Option<bool>,
    pub supports_storage_encryption: Option<bool>,
    pub vpc: Option<bool>,
}

Contains a list of available options for a DB instance.

This data type is used as a response element in the DescribeOrderableDBInstanceOptions action.

Fields
availability_zones: Option<Vec<AvailabilityZone>>

A list of Availability Zones for a DB instance.

db_instance_class: Option<String>

The DB instance class for a DB instance.

engine: Option<String>

The engine type of a DB instance.

engine_version: Option<String>

The engine version of a DB instance.

license_model: Option<String>

The license model for a DB instance.

max_iops_per_db_instance: Option<i64>

Maximum total provisioned IOPS for a DB instance.

max_iops_per_gib: Option<f64>

Maximum provisioned IOPS per GiB for a DB instance.

max_storage_size: Option<i64>

Maximum storage size for a DB instance.

min_iops_per_db_instance: Option<i64>

Minimum total provisioned IOPS for a DB instance.

min_iops_per_gib: Option<f64>

Minimum provisioned IOPS per GiB for a DB instance.

min_storage_size: Option<i64>

Minimum storage size for a DB instance.

multi_az_capable: Option<bool>

Indicates whether a DB instance is Multi-AZ capable.

read_replica_capable: Option<bool>

Indicates whether a DB instance can have a Read Replica.

storage_type: Option<String>

Indicates the storage type for a DB instance.

supports_enhanced_monitoring: Option<bool>

Indicates whether a DB instance supports Enhanced Monitoring at intervals from 1 to 60 seconds.

supports_iam_database_authentication: Option<bool>

Indicates whether a DB instance supports IAM database authentication.

supports_iops: Option<bool>

Indicates whether a DB instance supports provisioned IOPS.

supports_performance_insights: Option<bool>

True if a DB instance supports Performance Insights, otherwise false.

supports_storage_encryption: Option<bool>

Indicates whether a DB instance supports encrypted storage.

vpc: Option<bool>

Indicates whether a DB instance is in a VPC.