Struct rusoto_rds::RDSOption

pub struct RDSOption {
    pub db_security_group_memberships: Option<Vec<DBSecurityGroupMembership>>,
    pub option_description: Option<String>,
    pub option_name: Option<String>,
    pub option_settings: Option<Vec<OptionSetting>>,
    pub option_version: Option<String>,
    pub permanent: Option<bool>,
    pub persistent: Option<bool>,
    pub port: Option<i64>,
    pub vpc_security_group_memberships: Option<Vec<VpcSecurityGroupMembership>>,
}

Option details.

Fields
db_security_group_memberships: Option<Vec<DBSecurityGroupMembership>>

If the option requires access to a port, then this DB security group allows access to the port.

option_description: Option<String>

The description of the option.

option_name: Option<String>

The name of the option.

option_settings: Option<Vec<OptionSetting>>

The option settings for this option.

option_version: Option<String>

The version of the option.

permanent: Option<bool>

Indicate if this option is permanent.

persistent: Option<bool>

Indicate if this option is persistent.

port: Option<i64>

If required, the port configured for this option to use.

vpc_security_group_memberships: Option<Vec<VpcSecurityGroupMembership>>

If the option requires access to a port, then this VPC security group allows access to the port.