Struct rusoto_rds::RdsClient

pub struct RdsClient { /* fields omitted */ }

A client for the Amazon RDS API.

Methods

impl RdsClient

pub fn new(region: Region) -> RdsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> RdsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,