Struct rusoto_rds::RestoreDBClusterFromS3Message

pub struct RestoreDBClusterFromS3Message {
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub db_cluster_identifier: String,
    pub db_cluster_parameter_group_name: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub database_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub kms_key_id: Option<String>,
    pub master_user_password: String,
    pub master_username: String,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub s3_bucket_name: String,
    pub s3_ingestion_role_arn: String,
    pub s3_prefix: Option<String>,
    pub source_engine: String,
    pub source_engine_version: String,
    pub storage_encrypted: Option<bool>,
    pub tags: Option<Vec<Tag>>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
availability_zones: Option<Vec<String>>

A list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

backup_retention_period: Option<i64>

The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

character_set_name: Option<String>

A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

db_cluster_identifier: String

The name of the DB cluster to create from the source data in the S3 bucket. This parameter is isn't case-sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

db_cluster_parameter_group_name: Option<String>

The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 will be used.

Constraints:

Must be 1 to 255 alphanumeric characters

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

db_subnet_group_name: Option<String>

A DB subnet group to associate with the restored DB cluster.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

database_name: Option<String>

The database name for the restored DB cluster.

enable_iam_database_authentication: Option<bool>

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

engine: String

The name of the database engine to be used for the restored DB cluster.

Valid Values: aurora

engine_version: Option<String>

The version number of the database engine to use.

Aurora

Example: 5.6.10a

kms_key_id: Option<String>

The KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KM encryption key.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

master_user_password: String

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

master_username: String

The name of the master user for the restored DB cluster.

Constraints:

Must be 1 to 16 alphanumeric characters.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

option_group_name: Option<String>

A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options cannot be removed from an option group. An option group cannot be removed from a DB cluster once it is associated with a DB cluster.

port: Option<i64>

The port number on which the instances in the restored DB cluster accept connections.

Default: 3306

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

Default: A 30-minute window selected at random from an 8-hour block of time per region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Times should be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

s3_bucket_name: String

The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

s3_ingestion_role_arn: String

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

s3_prefix: Option<String>

The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

source_engine: String

The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: mysql

source_engine_version: String

The version of the database that the backup files were created from.

MySQL version 5.5 and 5.6 are supported.

Example: 5.6.22

storage_encrypted: Option<bool>

Specifies whether the restored DB cluster is encrypted.

tags: Option<Vec<Tag>> vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with the restored DB cluster.