Struct rusoto_rds::RestoreDBClusterFromSnapshotMessage

pub struct RestoreDBClusterFromSnapshotMessage {
    pub availability_zones: Option<Vec<String>>,
    pub db_cluster_identifier: String,
    pub db_subnet_group_name: Option<String>,
    pub database_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub kms_key_id: Option<String>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub snapshot_identifier: String,
    pub tags: Option<Vec<Tag>>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
availability_zones: Option<Vec<String>>

Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

db_cluster_identifier: String

The name of the DB cluster to create from the DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

db_subnet_group_name: Option<String>

The name of the DB subnet group to use for the new DB cluster.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

database_name: Option<String>

The database name for the restored DB cluster.

enable_iam_database_authentication: Option<bool>

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

engine: String

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

engine_version: Option<String>

The version of the database engine to use for the new DB cluster.

kms_key_id: Option<String>

The KMS key identifier to use when restoring an encrypted DB cluster from a DB cluster snapshot.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If you do not specify a value for the KmsKeyId parameter, then the following will occur:

If the DB cluster snapshot is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB cluster snapshot.

If the DB cluster snapshot is not encrypted, then the restored DB cluster is encrypted using the specified encryption key.

option_group_name: Option<String>

The name of the option group to use for the restored DB cluster.

port: Option<i64>

The port number on which the new DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

snapshot_identifier: String

The identifier for the DB cluster snapshot to restore from.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

tags: Option<Vec<Tag>>

The tags to be assigned to the restored DB cluster.

vpc_security_group_ids: Option<Vec<String>>

A list of VPC security groups that the new DB cluster will belong to.