Struct rusoto_rds::RestoreDBClusterToPointInTimeMessage

pub struct RestoreDBClusterToPointInTimeMessage {
    pub db_cluster_identifier: String,
    pub db_subnet_group_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub kms_key_id: Option<String>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub restore_to_time: Option<String>,
    pub restore_type: Option<String>,
    pub source_db_cluster_identifier: String,
    pub tags: Option<Vec<Tag>>,
    pub use_latest_restorable_time: Option<bool>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
db_cluster_identifier: String

The name of the new DB cluster to be created.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

db_subnet_group_name: Option<String>

The DB subnet group name to use for the new DB cluster.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

enable_iam_database_authentication: Option<bool>

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

kms_key_id: Option<String>

The KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

You can restore to a new DB cluster and encrypt the new DB cluster with a KMS key that is different than the KMS key used to encrypt the source DB cluster. The new DB cluster will be encrypted with the KMS key identified by the KmsKeyId parameter.

If you do not specify a value for the KmsKeyId parameter, then the following will occur:

If the DB cluster is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the source DB cluster.

If the DB cluster is not encrypted, then the restored DB cluster is not encrypted.

If DBClusterIdentifier refers to a DB cluster that is not encrypted, then the restore request is rejected.

option_group_name: Option<String>

The name of the option group for the new DB cluster.

port: Option<i64>

The port number on which the new DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

restore_to_time: Option<String>

The date and time to restore the DB cluster to.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Must be specified if UseLatestRestorableTime parameter is not provided

Cannot be specified if UseLatestRestorableTime parameter is true

Cannot be specified if RestoreType parameter is copy-on-write

Example: 2015-03-07T23:45:00Z

restore_type: Option<String>

The type of restore to be performed. You can specify one of the following values:

full-copy - The new DB cluster is restored as a full copy of the source DB cluster.

copy-on-write - The new DB cluster is restored as a clone of the source DB cluster.

Constraints: You cannot specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11.

If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

source_db_cluster_identifier: String

The identifier of the source DB cluster from which to restore.

Constraints:

Must be the identifier of an existing database instance

Must contain from 1 to 63 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

tags: Option<Vec<Tag>> use_latest_restorable_time: Option<bool>

A value that is set to true to restore the DB cluster to the latest restorable backup time, and false otherwise.

Default: false

Constraints: Cannot be specified if RestoreToTime parameter is provided.

vpc_security_group_ids: Option<Vec<String>>

A list of VPC security groups that the new DB cluster belongs to.