Struct rusoto_rds::RestoreDBInstanceFromDBSnapshotMessage

pub struct RestoreDBInstanceFromDBSnapshotMessage {
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_instance_class: Option<String>,
    pub db_instance_identifier: String,
    pub db_name: Option<String>,
    pub db_snapshot_identifier: String,
    pub db_subnet_group_name: Option<String>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: Option<String>,
    pub iops: Option<i64>,
    pub license_model: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub publicly_accessible: Option<bool>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
}

Fields
auto_minor_version_upgrade: Option<bool>

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

availability_zone: Option<String>

The EC2 Availability Zone that the database instance will be created in.

Default: A random, system-chosen Availability Zone.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

Example: us-east-1a

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the restored DB instance to snapshots of the DB instance; otherwise false. The default is false.

db_instance_class: Option<String>

The compute and memory capacity of the Amazon RDS DB instance.

Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large

db_instance_identifier: String

Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens (1 to 15 for SQL Server)

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

db_name: Option<String>

The database name for the restored DB instance.

This parameter doesn't apply to the MySQL, PostgreSQL, or MariaDB engines.

db_snapshot_identifier: String

The identifier for the DB snapshot to restore from.

Constraints:

Must contain from 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

db_subnet_group_name: Option<String>

The DB subnet group name to use for the new instance.

Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

domain: Option<String>

Specify the Active Directory Domain to restore the instance in.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_iam_database_authentication: Option<bool>

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false.

You can enable IAM database authentication for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher.

Default: false

engine: Option<String>

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source. You can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot.

Valid Values: MySQL | mariadb | oracle-se1 | oracle-se | oracle-ee | sqlserver-ee | sqlserver-se | sqlserver-ex | sqlserver-web | postgres | aurora

iops: Option<i64>

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter is not specified, the IOPS value will be taken from the backup. If this parameter is set to 0, the new instance will be converted to a non-PIOPS instance, which will take additional time, though your DB instance will be available for connections before the conversion starts.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine is not supported.

license_model: Option<String>

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

option_group_name: Option<String>

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

port: Option<i64>

The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be 1150-65535

publicly_accessible: Option<bool>

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC: true

VPC: false

If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

tags: Option<Vec<Tag>> tde_credential_arn: Option<String>

The ARN from the Key Store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the Key Store in order to access the device.