Struct rusoto_rds::RestoreDBInstanceFromS3Message

pub struct RestoreDBInstanceFromS3Message {
    pub allocated_storage: Option<i64>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub backup_retention_period: Option<i64>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_instance_class: String,
    pub db_instance_identifier: String,
    pub db_name: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub db_subnet_group_name: Option<String>,
    pub deletion_protection: Option<bool>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_performance_insights: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub license_model: Option<String>,
    pub master_user_password: Option<String>,
    pub master_username: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub performance_insights_kms_key_id: Option<String>,
    pub performance_insights_retention_period: Option<i64>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub processor_features: Option<Vec<ProcessorFeature>>,
    pub publicly_accessible: Option<bool>,
    pub s3_bucket_name: String,
    pub s3_ingestion_role_arn: String,
    pub s3_prefix: Option<String>,
    pub source_engine: String,
    pub source_engine_version: String,
    pub storage_encrypted: Option<bool>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub use_default_processor_features: Option<bool>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
allocated_storage: Option<i64>

The amount of storage (in gigabytes) to allocate initially for the DB instance. Follow the allocation rules specified in CreateDBInstance.

Be sure to allocate enough memory for your new DB instance so that the restore operation can succeed. You can also allocate additional memory for future growth.

auto_minor_version_upgrade: Option<bool>

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are not applied automatically.

availability_zone: Option<String>

The Availability Zone that the DB instance is created in. For information about AWS Regions and Availability Zones, see Regions and Availability Zones in the Amazon RDS User Guide.

Default: A random, system-chosen Availability Zone in the endpoint's AWS Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same AWS Region as the current endpoint.

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. For more information, see CreateDBInstance.

copy_tags_to_snapshot: Option<bool>

A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

db_instance_class: String

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Importing from Amazon S3 is not supported on the db.t2.micro DB instance class.

db_instance_identifier: String

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

db_name: Option<String>

The name of the database to create when the DB instance is created. Follow the naming rules specified in CreateDBInstance.

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default parameter group for the specified engine is used.

db_security_groups: Option<Vec<String>>

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB instance.

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

enable_iam_database_authentication: Option<bool>

A value that indicates whether to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

enable_performance_insights: Option<bool>

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

engine: String

The name of the database engine to be used for this instance.

Valid Values: mysql

engine_version: Option<String>

The version number of the database engine to use. Choose the latest minor version of your database engine. For information about engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to allocate initially for the DB instance. For information about valid Iops values, see see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

kms_key_id: Option<String>

The AWS KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same AWS account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

license_model: Option<String>

The license model for this DB instance. Use general-public-license.

master_user_password: Option<String>

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

master_username: Option<String>

The name for the master user.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

Default: 0

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, see Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

multi_az: Option<bool>

A value that indicates whether the DB instance is a Multi-AZ deployment. If the DB instance is a Multi-AZ deployment, you can't set the AvailabilityZone parameter.

option_group_name: Option<String>

The name of the option group to associate with this DB instance. If this argument is omitted, the default option group for the specified engine is used.

performance_insights_kms_key_id: Option<String>

The AWS KMS key identifier for encryption of Performance Insights data. The KMS key ID is the Amazon Resource Name (ARN), the KMS key identifier, or the KMS key alias for the KMS encryption key.

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

performance_insights_retention_period: Option<i64>

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

port: Option<i64>

The port number on which the database accepts connections.

Type: Integer

Valid Values: 1150-65535

Default: 3306

preferred_backup_window: Option<String>

The time range each day during which automated backups are created if automated backups are enabled. For more information, see The Backup Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format ddd:hh24:mi-ddd:hh24:mi.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred backup window.

Must be at least 30 minutes.

processor_features: Option<Vec<ProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

publicly_accessible: Option<bool>

A value that indicates whether the DB instance is publicly accessible. When the DB instance is publicly accessible, it is an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. When the DB instance is not publicly accessible, it is an internal instance with a DNS name that resolves to a private IP address. For more information, see CreateDBInstance.

s3_bucket_name: String

The name of your Amazon S3 bucket that contains your database backup file.

s3_ingestion_role_arn: String

An AWS Identity and Access Management (IAM) role to allow Amazon RDS to access your Amazon S3 bucket.

s3_prefix: Option<String>

The prefix of your Amazon S3 bucket.

source_engine: String

The name of the engine of your source database.

Valid Values: mysql

source_engine_version: String

The engine version of your source database.

Valid Values: 5.6

storage_encrypted: Option<bool>

A value that indicates whether the new DB instance is encrypted or not.

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise gp2

tags: Option<Vec<Tag>>

A list of tags to associate with this DB instance. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

use_default_processor_features: Option<bool>

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

vpc_security_group_ids: Option<Vec<String>>

A list of VPC security groups to associate with this DB instance.