Struct rusoto_rds::RestoreDBInstanceToPointInTimeMessage

pub struct RestoreDBInstanceToPointInTimeMessage {
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_instance_class: Option<String>,
    pub db_name: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub deletion_protection: Option<bool>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub enable_iam_database_authentication: Option<bool>,
    pub engine: Option<String>,
    pub iops: Option<i64>,
    pub license_model: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub processor_features: Option<Vec<ProcessorFeature>>,
    pub publicly_accessible: Option<bool>,
    pub restore_time: Option<String>,
    pub source_db_instance_identifier: Option<String>,
    pub source_dbi_resource_id: Option<String>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub target_db_instance_identifier: String,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub use_default_processor_features: Option<bool>,
    pub use_latest_restorable_time: Option<bool>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
auto_minor_version_upgrade: Option<bool>

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

availability_zone: Option<String>

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

copy_tags_to_snapshot: Option<bool>

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

db_instance_class: Option<String>

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

db_name: Option<String>

The database name for the restored DB instance.

This parameter is not used for the MySQL or MariaDB engines.

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

db_subnet_group_name: Option<String>

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

domain: Option<String>

Specify the Active Directory Domain to restore the instance in.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

enable_iam_database_authentication: Option<bool>

A value that indicates whether to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

You can enable IAM database authentication for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

engine: Option<String>

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-se2

oracle-se1

oracle-se

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine is not supported.

license_model: Option<String>

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

multi_az: Option<bool>

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

option_group_name: Option<String>

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

port: Option<i64>

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

processor_features: Option<Vec<ProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

publicly_accessible: Option<bool>

A value that indicates whether the DB instance is publicly accessible. When the DB instance is publicly accessible, it is an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. When the DB instance is not publicly accessible, it is an internal instance with a DNS name that resolves to a private IP address. For more information, see CreateDBInstance.

restore_time: Option<String>

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

source_db_instance_identifier: Option<String>

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

source_dbi_resource_id: Option<String>

The resource ID of the source DB instance from which to restore.

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

tags: Option<Vec<Tag>>target_db_instance_identifier: String

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

use_default_processor_features: Option<bool>

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

use_latest_restorable_time: Option<bool>

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance is not restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.