Enum rusoto_rds::RestoreDBClusterFromS3Error

pub enum RestoreDBClusterFromS3Error {
    DBClusterAlreadyExistsFault(String),
    DBClusterNotFoundFault(String),
    DBClusterParameterGroupNotFoundFault(String),
    DBClusterQuotaExceededFault(String),
    DBSubnetGroupNotFoundFault(String),
    InsufficientStorageClusterCapacityFault(String),
    InvalidDBClusterStateFault(String),
    InvalidDBSubnetGroupStateFault(String),
    InvalidS3BucketFault(String),
    InvalidSubnet(String),
    InvalidVPCNetworkStateFault(String),
    KMSKeyNotAccessibleFault(String),
    StorageQuotaExceededFault(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by RestoreDBClusterFromS3

Variants
DBClusterAlreadyExistsFault(String)

User already has a DB cluster with the given identifier.

DBClusterNotFoundFault(String)

DBClusterIdentifier does not refer to an existing DB cluster.

DBClusterParameterGroupNotFoundFault(String)

DBClusterParameterGroupName does not refer to an existing DB Cluster parameter group.

DBClusterQuotaExceededFault(String)

User attempted to create a new DB cluster and the user has already reached the maximum allowed DB cluster quota.

DBSubnetGroupNotFoundFault(String)

DBSubnetGroupName does not refer to an existing DB subnet group.

InsufficientStorageClusterCapacityFault(String)

There is insufficient storage available for the current action. You may be able to resolve this error by updating your subnet group to use different Availability Zones that have more storage available.

InvalidDBClusterStateFault(String)

The DB cluster is not in a valid state.

InvalidDBSubnetGroupStateFault(String)

The DB subnet group cannot be deleted because it is in use.

InvalidS3BucketFault(String)

The specified Amazon S3 bucket name could not be found or Amazon RDS is not authorized to access the specified Amazon S3 bucket. Verify the SourceS3BucketName and S3IngestionRoleArn values and try again.

InvalidSubnet(String)

The requested subnet is invalid, or multiple subnets were requested that are not all in a common VPC.

InvalidVPCNetworkStateFault(String)

DB subnet group does not cover all Availability Zones after it is created because users' change.

KMSKeyNotAccessibleFault(String)

Error accessing KMS key.

StorageQuotaExceededFault(String)

Request would result in user exceeding the allowed amount of storage available across all DB instances.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl RestoreDBClusterFromS3Error

pub fn from_body(body: &str) -> RestoreDBClusterFromS3Error