Struct rusoto_rds::CreateDBInstanceMessage

pub struct CreateDBInstanceMessage {
    pub allocated_storage: Option<i64>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_cluster_identifier: Option<String>,
    pub db_instance_class: String,
    pub db_instance_identifier: String,
    pub db_name: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub db_subnet_group_name: Option<String>,
    pub deletion_protection: Option<bool>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_performance_insights: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub license_model: Option<String>,
    pub master_user_password: Option<String>,
    pub master_username: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub performance_insights_kms_key_id: Option<String>,
    pub performance_insights_retention_period: Option<i64>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub processor_features: Option<Vec<ProcessorFeature>>,
    pub promotion_tier: Option<i64>,
    pub publicly_accessible: Option<bool>,
    pub storage_encrypted: Option<bool>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub timezone: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
allocated_storage: Option<i64>

The amount of storage (in gibibytes) to allocate for the DB instance.

Type: Integer

Amazon Aurora

Not applicable. Aurora cluster volumes automatically grow as the amount of data in your database increases, though you are only charged for the space that you use in an Aurora cluster volume.

MySQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

MariaDB

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

PostgreSQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

Oracle

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 10 to 3072.

SQL Server

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 100 to 16384.

Magnetic storage (standard):

Enterprise and Standard editions: Must be an integer from 200 to 1024.

Web and Express editions: Must be an integer from 20 to 1024.

auto_minor_version_upgrade: Option<bool>

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are applied automatically.

availability_zone: Option<String>

The Availability Zone (AZ) where the database will be created. For information on AWS Regions and Availability Zones, see Regions and Availability Zones.

Default: A random, system-chosen Availability Zone in the endpoint's AWS Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same AWS Region as the current endpoint.

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Amazon Aurora

Not applicable. The retention period for automated backups is managed by the DB cluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Can't be set to 0 if the DB instance is a source to Read Replicas

character_set_name: Option<String>

For supported engines, indicates that the DB instance should be associated with the specified CharacterSet.

Amazon Aurora

Not applicable. The character set is managed by the DB cluster. For more information, see CreateDBCluster.

copy_tags_to_snapshot: Option<bool>

A value that indicates whether to copy tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting.

db_cluster_identifier: Option<String>

The identifier of the DB cluster that the instance will belong to.

db_instance_class: String

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

db_instance_identifier: String

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use.

MySQL

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Can't be a word reserved by the specified database engine

MariaDB

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Can't be a word reserved by the specified database engine

PostgreSQL

The name of the database to create when the DB instance is created. If this parameter is not specified, the default "postgres" database is created in the DB instance.

Constraints:

Must contain 1 to 63 letters, numbers, or underscores.

Must begin with a letter or an underscore. Subsequent characters can be letters, underscores, or digits (0-9).

Can't be a word reserved by the specified database engine

Oracle

The Oracle System ID (SID) of the created DB instance. If you specify null, the default value ORCL is used. You can't specify the string NULL, or any other reserved word, for DBName.

Default: ORCL

Constraints:

Can't be longer than 8 characters

SQL Server

Not applicable. Must be null.

Amazon Aurora

The name of the database to create when the primary instance of the DB cluster is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Can't be a word reserved by the specified database engine

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

db_security_groups: Option<Vec<String>>

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

domain: Option<String>

For an Amazon RDS DB instance that's running Microsoft SQL Server, this parameter specifies the Active Directory directory ID to create the instance in. Amazon RDS uses Windows Authentication to authenticate users that connect to the DB instance. For more information, see Using Windows Authentication with an Amazon RDS DB Instance Running Microsoft SQL Server in the Amazon RDS User Guide.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of log types that need to be enabled for exporting to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Relational Database Service User Guide.

enable_iam_database_authentication: Option<bool>

A value that indicates whether to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

You can enable IAM database authentication for the following database engines:

Amazon Aurora

Not applicable. Mapping AWS IAM accounts to database accounts is managed by the DB cluster.

MySQL

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

enable_performance_insights: Option<bool>

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

engine: String

The name of the database engine to be used for this instance.

Not every database engine is available for every AWS Region.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-se2

oracle-se1

oracle-se

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

engine_version: Option<String>

The version number of the database engine to use.

For a list of valid engine versions, use the DescribeDBEngineVersions action.

The following are the database engines and links to information about the major and minor versions that are available with Amazon RDS. Not every database engine is available for every AWS Region.

Amazon Aurora

Not applicable. The version number of the database engine to be used by the DB instance is managed by the DB cluster.

MariaDB

See MariaDB on Amazon RDS Versions in the Amazon RDS User Guide.

Microsoft SQL Server

See Version and Feature Support on Amazon RDS in the Amazon RDS User Guide.

MySQL

See MySQL on Amazon RDS Versions in the Amazon RDS User Guide.

Oracle

See Oracle Database Engine Release Notes in the Amazon RDS User Guide.

PostgreSQL

See Supported PostgreSQL Database Versions in the Amazon RDS User Guide.

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance. For information about valid Iops values, see see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: Must be a multiple between 1 and 50 of the storage amount for the DB instance.

kms_key_id: Option<String>

The AWS KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same AWS account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Amazon Aurora

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If StorageEncrypted is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

license_model: Option<String>

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

master_user_password: Option<String>

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Amazon Aurora

Not applicable. The password for the master user is managed by the DB cluster.

MariaDB

Constraints: Must contain from 8 to 41 characters.

Microsoft SQL Server

Constraints: Must contain from 8 to 128 characters.

MySQL

Constraints: Must contain from 8 to 41 characters.

Oracle

Constraints: Must contain from 8 to 30 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

master_username: Option<String>

The name for the master user.

Amazon Aurora

Not applicable. The name for the master user is managed by the DB cluster.

MariaDB

Constraints:

Required for MariaDB.

Must be 1 to 16 letters or numbers.

Can't be a reserved word for the chosen database engine.

Microsoft SQL Server

Constraints:

Required for SQL Server.

Must be 1 to 128 letters or numbers.

The first character must be a letter.

Can't be a reserved word for the chosen database engine.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

Oracle

Constraints:

Required for Oracle.

Must be 1 to 30 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

multi_az: Option<bool>

A value that indicates whether the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

option_group_name: Option<String>

Indicates that the DB instance should be associated with the specified option group.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

performance_insights_kms_key_id: Option<String>

The AWS KMS key identifier for encryption of Performance Insights data. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

performance_insights_retention_period: Option<i64>

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

port: Option<i64>

The port number on which the database accepts connections.

MySQL

Default: 3306

Valid Values: 1150-65535

Type: Integer

MariaDB

Default: 3306

Valid Values: 1150-65535

Type: Integer

PostgreSQL

Default: 5432

Valid Values: 1150-65535

Type: Integer

Oracle

Default: 1521

Valid Values: 1150-65535

SQL Server

Default: 1433

Valid Values: 1150-65535 except for 1434, 3389, 47001, 49152, and 49152 through 49156.

Amazon Aurora

Default: 3306

Valid Values: 1150-65535

Type: Integer

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. For more information, see The Backup Window in the Amazon RDS User Guide.

Amazon Aurora

Not applicable. The daily time range for creating automated backups is managed by the DB cluster.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. To see the time blocks available, see Adjusting the Preferred DB Instance Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window.

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

processor_features: Option<Vec<ProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

promotion_tier: Option<i64>

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

Default: 1

Valid Values: 0 - 15

publicly_accessible: Option<bool>

A value that indicates whether the DB instance is publicly accessible. When the DB instance is publicly accessible, it is an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. When the DB instance is not publicly accessible, it is an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether DBSubnetGroupName is specified.

If DBSubnetGroupName is not specified, and PubliclyAccessible is not specified, the following applies:

If the default VPC in the target region doesn’t have an Internet gateway attached to it, the DB instance is private.

If the default VPC in the target region has an Internet gateway attached to it, the DB instance is public.

If DBSubnetGroupName is specified, and PubliclyAccessible is not specified, the following applies:

If the subnets are part of a VPC that doesn’t have an Internet gateway attached to it, the DB instance is private.

If the subnets are part of a VPC that has an Internet gateway attached to it, the DB instance is public.

storage_encrypted: Option<bool>

A value that indicates whether the DB instance is encrypted. By default, it is not encrypted.

Amazon Aurora

Not applicable. The encryption for DB instances is managed by the DB cluster.

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

tags: Option<Vec<Tag>>

Tags to assign to the DB instance.

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

timezone: Option<String>

The time zone of the DB instance. The time zone parameter is currently supported only by Microsoft SQL Server.

vpc_security_group_ids: Option<Vec<String>>

A list of Amazon EC2 VPC security groups to associate with this DB instance.

Amazon Aurora

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.