Struct rusoto_rds::CreateDBInstanceMessage

pub struct CreateDBInstanceMessage {
    pub allocated_storage: Option<i64>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub availability_zone: Option<String>,
    pub backup_retention_period: Option<i64>,
    pub character_set_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub db_cluster_identifier: Option<String>,
    pub db_instance_class: String,
    pub db_instance_identifier: String,
    pub db_name: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub db_subnet_group_name: Option<String>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_performance_insights: Option<bool>,
    pub engine: String,
    pub engine_version: Option<String>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub license_model: Option<String>,
    pub master_user_password: Option<String>,
    pub master_username: Option<String>,
    pub monitoring_interval: Option<i64>,
    pub monitoring_role_arn: Option<String>,
    pub multi_az: Option<bool>,
    pub option_group_name: Option<String>,
    pub performance_insights_kms_key_id: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub promotion_tier: Option<i64>,
    pub publicly_accessible: Option<bool>,
    pub storage_encrypted: Option<bool>,
    pub storage_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub timezone: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
allocated_storage: Option<i64>

The amount of storage (in gibibytes) to allocate for the DB instance.

Type: Integer

Amazon Aurora

Not applicable. Aurora cluster volumes automatically grow as the amount of data in your database increases, though you are only charged for the space that you use in an Aurora cluster volume.

MySQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1): Must be an integer from 100 to 16384.

Magnetic storage (standard): Must be an integer from 5 to 3072.

MariaDB

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1): Must be an integer from 100 to 16384.

Magnetic storage (standard): Must be an integer from 5 to 3072.

PostgreSQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1): Must be an integer from 100 to 16384.

Magnetic storage (standard): Must be an integer from 5 to 3072.

Oracle

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1): Must be an integer from 100 to 16384.

Magnetic storage (standard): Must be an integer from 10 to 3072.

SQL Server

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 100 to 16384.

Magnetic storage (standard):

Enterprise and Standard editions: Must be an integer from 200 to 1024.

Web and Express editions: Must be an integer from 20 to 1024.

auto_minor_version_upgrade: Option<bool>

Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

availability_zone: Option<String>

The EC2 Availability Zone that the DB instance is created in. For information on AWS Regions and Availability Zones, see Regions and Availability Zones.

Default: A random, system-chosen Availability Zone in the endpoint's AWS Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same AWS Region as the current endpoint.

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Amazon Aurora

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

character_set_name: Option<String>

For supported engines, indicates that the DB instance should be associated with the specified CharacterSet.

Amazon Aurora

Not applicable. The character set is managed by the DB cluster. For more information, see CreateDBCluster.

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

db_cluster_identifier: Option<String>

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

db_instance_class: String

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

db_instance_identifier: String

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use.

Type: String

MySQL

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Cannot be a word reserved by the specified database engine

MariaDB

The name of the database to create when the DB instance is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Cannot be a word reserved by the specified database engine

PostgreSQL

The name of the database to create when the DB instance is created. If this parameter is not specified, the default "postgres" database is created in the DB instance.

Constraints:

Must contain 1 to 63 letters, numbers, or underscores.

Must begin with a letter or an underscore. Subsequent characters can be letters, underscores, or digits (0-9).

Cannot be a word reserved by the specified database engine

Oracle

The Oracle System ID (SID) of the created DB instance. If you specify null, the default value ORCL is used. You can't specify the string NULL, or any other reserved word, for DBName.

Default: ORCL

Constraints:

Cannot be longer than 8 characters

SQL Server

Not applicable. Must be null.

Amazon Aurora

The name of the database to create when the primary instance of the DB cluster is created. If this parameter is not specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Cannot be a word reserved by the specified database engine

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

db_security_groups: Option<Vec<String>>

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

domain: Option<String>

Specify the Active Directory Domain to create the instance in.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of log types that need to be enabled for exporting to CloudWatch Logs.

enable_iam_database_authentication: Option<bool>

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines:

Amazon Aurora

Not applicable. Mapping AWS IAM accounts to database accounts is managed by the DB cluster. For more information, see CreateDBCluster.

MySQL

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Default: false

enable_performance_insights: Option<bool>

True to enable Performance Insights for the DB instance, and otherwise false.

engine: String

The name of the database engine to be used for this instance.

Not every database engine is available for every AWS Region.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-se2

oracle-se1

oracle-se

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

engine_version: Option<String>

The version number of the database engine to use.

The following are the database engines and major and minor versions that are available with Amazon RDS. Not every database engine is available for every AWS Region.

Amazon Aurora

Not applicable. The version number of the database engine to be used by the DB instance is managed by the DB cluster. For more information, see CreateDBCluster.

MariaDB

10.2.12 (supported in all AWS Regions)

10.2.11 (supported in all AWS Regions)

10.1.31 (supported in all AWS Regions)

10.1.26 (supported in all AWS Regions)

10.1.23 (supported in all AWS Regions)

10.1.19 (supported in all AWS Regions)

10.1.14 (supported in all AWS Regions except us-east-2)

10.0.34 (supported in all AWS Regions)

10.0.32 (supported in all AWS Regions)

10.0.31 (supported in all AWS Regions)

10.0.28 (supported in all AWS Regions)

10.0.24 (supported in all AWS Regions)

10.0.17 (supported in all AWS Regions except us-east-2, ca-central-1, eu-west-2)

Microsoft SQL Server 2017

14.00.1000.169.v1 (supported for all editions, and all AWS Regions)

Microsoft SQL Server 2016

13.00.4451.0.v1 (supported for all editions, and all AWS Regions)

13.00.4422.0.v1 (supported for all editions, and all AWS Regions)

13.00.2164.0.v1 (supported for all editions, and all AWS Regions)

Microsoft SQL Server 2014

12.00.5546.0.v1 (supported for all editions, and all AWS Regions)

12.00.5000.0.v1 (supported for all editions, and all AWS Regions)

12.00.4422.0.v1 (supported for all editions except Enterprise Edition, and all AWS Regions except ca-central-1 and eu-west-2)

Microsoft SQL Server 2012

11.00.6594.0.v1 (supported for all editions, and all AWS Regions)

11.00.6020.0.v1 (supported for all editions, and all AWS Regions)

11.00.5058.0.v1 (supported for all editions, and all AWS Regions except us-east-2, ca-central-1, and eu-west-2)

11.00.2100.60.v1 (supported for all editions, and all AWS Regions except us-east-2, ca-central-1, and eu-west-2)

Microsoft SQL Server 2008 R2

10.50.6529.0.v1 (supported for all editions, and all AWS Regions except us-east-2, ca-central-1, and eu-west-2)

10.50.6000.34.v1 (supported for all editions, and all AWS Regions except us-east-2, ca-central-1, and eu-west-2)

10.50.2789.0.v1 (supported for all editions, and all AWS Regions except us-east-2, ca-central-1, and eu-west-2)

MySQL

5.7.21 (supported in all AWS regions)

5.7.19 (supported in all AWS regions)

5.7.17 (supported in all AWS regions)

5.7.16 (supported in all AWS regions)

5.6.39 (supported in all AWS Regions)

5.6.37 (supported in all AWS Regions)

5.6.35 (supported in all AWS Regions)

5.6.34 (supported in all AWS Regions)

5.6.29 (supported in all AWS Regions)

5.6.27 (supported in all AWS Regions except us-east-2, ca-central-1, eu-west-2)

5.5.59 (supported in all AWS Regions)

5.5.57 (supported in all AWS Regions)

5.5.54 (supported in all AWS Regions)

5.5.53 (supported in all AWS Regions)

5.5.46 (supported in all AWS Regions)

Oracle 12c

12.1.0.2.v9 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v8 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v7 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v6 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v5 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v4 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v3 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v2 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

12.1.0.2.v1 (supported for EE in all AWS regions, and SE2 in all AWS regions except us-gov-west-1)

Oracle 11g

11.2.0.4.v13 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v12 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v11 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v10 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v9 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v8 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v7 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v6 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v5 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v4 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v3 (supported for EE, SE1, and SE, in all AWS regions)

11.2.0.4.v1 (supported for EE, SE1, and SE, in all AWS regions)

PostgreSQL

Version 10.1

Version 9.6.x: 9.6.6 | 9.6.5 | 9.6.3 | 9.6.2 | 9.6.1

Version 9.5.x: 9.5.9 | 9.5.7 | 9.5.6 | 9.5.4 | 9.5.2

Version 9.4.x: 9.4.14 | 9.4.12 | 9.4.11 | 9.4.9 | 9.4.7

Version 9.3.x: 9.3.19 | 9.3.17 | 9.3.16 | 9.3.14 | 9.3.12

iops: Option<i64>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance. For information about valid Iops values, see see Amazon RDS Provisioned IOPS Storage to Improve Performance.

Constraints: Must be a multiple between 1 and 50 of the storage amount for the DB instance. Must also be an integer multiple of 1000. For example, if the size of your DB instance is 500 GiB, then your Iops value can be 2000, 3000, 4000, or 5000.

kms_key_id: Option<String>

The AWS KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same AWS account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Amazon Aurora

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

license_model: Option<String>

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

master_user_password: Option<String>

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Amazon Aurora

Not applicable. The password for the master user is managed by the DB cluster. For more information, see CreateDBCluster.

MariaDB

Constraints: Must contain from 8 to 41 characters.

Microsoft SQL Server

Constraints: Must contain from 8 to 128 characters.

MySQL

Constraints: Must contain from 8 to 41 characters.

Oracle

Constraints: Must contain from 8 to 30 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

master_username: Option<String>

The name for the master user.

Amazon Aurora

Not applicable. The name for the master user is managed by the DB cluster. For more information, see CreateDBCluster.

MariaDB

Constraints:

Required for MariaDB.

Must be 1 to 16 letters or numbers.

Cannot be a reserved word for the chosen database engine.

Microsoft SQL Server

Constraints:

Required for SQL Server.

Must be 1 to 128 letters or numbers.

The first character must be a letter.

Cannot be a reserved word for the chosen database engine.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

Oracle

Constraints:

Required for Oracle.

Must be 1 to 30 letters or numbers.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers.

First character must be a letter.

Cannot be a reserved word for the chosen database engine.

monitoring_interval: Option<i64>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to Setting Up and Enabling Enhanced Monitoring.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

option_group_name: Option<String>

Indicates that the DB instance should be associated with the specified option group.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

performance_insights_kms_key_id: Option<String>

The AWS KMS key identifier for encryption of Performance Insights data. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

port: Option<i64>

The port number on which the database accepts connections.

MySQL

Default: 3306

Valid Values: 1150-65535

Type: Integer

MariaDB

Default: 3306

Valid Values: 1150-65535

Type: Integer

PostgreSQL

Default: 5432

Valid Values: 1150-65535

Type: Integer

Oracle

Default: 1521

Valid Values: 1150-65535

SQL Server

Default: 1433

Valid Values: 1150-65535 except for 1434, 3389, 47001, 49152, and 49152 through 49156.

Amazon Aurora

Default: 3306

Valid Values: 1150-65535

Type: Integer

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. For more information, see The Backup Window.

Amazon Aurora

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. To see the time blocks available, see Adjusting the Preferred DB Instance Maintenance Window.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window.

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

promotion_tier: Option<i64>

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster.

Default: 1

Valid Values: 0 - 15

publicly_accessible: Option<bool>

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC: true

VPC: false

If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance is publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance is private.

storage_encrypted: Option<bool>

Specifies whether the DB instance is encrypted.

Amazon Aurora

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise standard

tags: Option<Vec<Tag>> tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

timezone: Option<String>

The time zone of the DB instance. The time zone parameter is currently supported only by Microsoft SQL Server.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB instance.

Amazon Aurora

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.