Struct rusoto_rds::ModifyDBClusterMessage

pub struct ModifyDBClusterMessage {
    pub apply_immediately: Option<bool>,
    pub backup_retention_period: Option<i64>,
    pub db_cluster_identifier: String,
    pub db_cluster_parameter_group_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub master_user_password: Option<String>,
    pub new_db_cluster_identifier: Option<String>,
    pub option_group_name: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
apply_immediately: Option<bool>

A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects the NewDBClusterIdentifier and MasterUserPassword values. If you set the ApplyImmediately parameter value to false, then changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

backup_retention_period: Option<i64>

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

db_cluster_identifier: String

The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

db_cluster_parameter_group_name: Option<String>

The name of the DB cluster parameter group to use for the DB cluster.

enable_iam_database_authentication: Option<bool>

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

master_user_password: Option<String>

The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

new_db_cluster_identifier: Option<String>

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

option_group_name: Option<String>

A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

port: Option<i64>

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

preferred_maintenance_window: Option<String>

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

vpc_security_group_ids: Option<Vec<String>>

A list of VPC security groups that the DB cluster will belong to.